+++ internal/logger/logger.go
@@ -13,10 +13,13 @@ const DefaultTimeLayout = "2006-01-02 15:04:05"
 // DefaultLineFormat is format to log line record.
 const DefaultLineFormat = "[%s] %s: %s\n%s\n\n"
 
+// ErrEmptyFileName is returned when trying to open file with empty name.
+var ErrEmptyFileName = errors.New("empty file name")
+
 // OpenFile opens file for append strings. Creates file if file not exist.
 func OpenFile(name string) (file *os.File, err error) {
 	if name == "" {
-		return nil, errors.New("empty file name")
+		return nil, ErrEmptyFileName
 	}
 
 	switch _, err = os.Stat(name); {