+++ internal/proto/websocket/websocket.go
@@ -1,16 +1,12 @@
 package websocket
 
-import (
-	"errors"
-
-	"github.com/gorcon/websocket"
-)
+import "github.com/gorcon/websocket"
 
 // Execute sends command to Execute to the remote server and returns
 // the response.
 func Execute(address string, password string, command string) (string, error) {
 	if command == "" {
-		return "", errors.New("command is not set")
+		return "", websocket.ErrCommandEmpty
 	}
 
 	console, err := websocket.Dial(address, password)